Fixed Crouch Blocking...
Albeit has shown I have to find a better approach to sprite extractions. One that will use a master list so an animation can pull any frame throughout the game and not just ones ripped in a sequence. This was found after requesting the "unblocking while crouched" animation as the last frame is pulled from the crouch animation which isn't within the "blocking while crouched" animation.